Include #3787
