+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/expiration/DccExpirationCheck.kt
@@ -1,79 +1,88 @@
 package de.rki.coronawarnapp.covidcertificate.expiration
 
 import de.rki.coronawarnapp.covidcertificate.common.certificate.CwaCovidCertificate
-import de.rki.coronawarnapp.covidcertificate.common.repository.CertificateContainerId
+import de.rki.coronawarnapp.covidcertificate.recovery.core.RecoveryCertificate
 import de.rki.coronawarnapp.covidcertificate.recovery.core.RecoveryCertificateRepository
+import de.rki.coronawarnapp.covidcertificate.vaccination.core.CovidCertificateSettings
+import de.rki.coronawarnapp.covidcertificate.vaccination.core.VaccinationCertificate
 import de.rki.coronawarnapp.covidcertificate.vaccination.core.repository.VaccinationRepository
+import de.rki.coronawarnapp.util.TimeStamper
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.sync.Mutex
+import kotlinx.coroutines.sync.withLock
+import org.joda.time.Duration
 import timber.log.Timber
 import javax.inject.Inject
+import javax.inject.Singleton
 
+@Singleton
 class DccExpirationCheck @Inject constructor(
-    private val dccExpirationNotification: DccExpirationNotification,
-    private val dccExpirationNotificationRepository: DccExpirationNotificationRepository,
+    private val dscCheckNotification: DccExpirationNotification,
     private val vaccinationRepository: VaccinationRepository,
-    private val recoveryCertificateRepository: RecoveryCertificateRepository,
+    private val recoveryRepository: RecoveryCertificateRepository,
+    private val covidCertificateSettings: CovidCertificateSettings,
+    private val timeStamper: TimeStamper,
 ) {
+    private val mutex = Mutex()
 
-    suspend fun checkExpirationStates() {
-        if (dccExpirationNotificationRepository.latestExecutionHasBeenToday()) {
-            Timber.i("Latest execution has been today. Abort.")
+    suspend fun checkStates() = mutex.withLock {
+        Timber.tag(TAG).v("checkStates()")
+
+        val lastCheck = covidCertificateSettings.lastDccStateBackgroundCheck.value
+        if (Duration(lastCheck, timeStamper.nowUTC).abs().standardDays < 1) {
+            Timber.tag(TAG).d("Last check was within 24h, skipping.")
             return
         }
 
-        val expiredList: MutableList<CertificateContainerId> = mutableListOf()
-        val expiringSoonList: MutableList<CertificateContainerId> = mutableListOf()
-        vaccinationRepository.vaccinationInfos.first().forEach {
-            val certificate = it.getMostRecentVaccinationCertificate
-            when (certificate.getState()) {
-                is CwaCovidCertificate.State.ExpiringSoon -> expiringSoonList.add(certificate.containerId)
-                is CwaCovidCertificate.State.Expired -> expiredList.add(certificate.containerId)
-                else -> { /*not relevant*/
+        val allCerts = getCertificates()
+
+        allCerts
+            .filter { it.getState() is CwaCovidCertificate.State.Expired }
+            .firstOrNull {
+                Timber.tag(TAG).w("Certificate expired: %s", it)
+                it.notifiedExpiredAt == null
+            }
+            ?.let {
+                if (dscCheckNotification.showExpiredNotification(it.containerId)) {
+                    setStateNotificationShown(it)
                 }
             }
-        }
-        recoveryCertificateRepository.certificates.first().forEach {
-            val certificate = it.recoveryCertificate
-            when (certificate.getState()) {
-                is CwaCovidCertificate.State.ExpiringSoon -> expiringSoonList.add(certificate.containerId)
-                is CwaCovidCertificate.State.Expired -> expiredList.add(certificate.containerId)
-                else -> { /*not relevant*/
+
+        allCerts
+            .filter { it.getState() is CwaCovidCertificate.State.ExpiringSoon }
+            .firstOrNull {
+                Timber.tag(TAG).w("Certificate expiring soon: %s", it)
+                it.notifiedExpiresSoonAt == null && it.notifiedExpiredAt == null
+            }
+            ?.let {
+                if (dscCheckNotification.showExpiresSoonNotification(it.containerId)) {
+                    setStateNotificationShown(it)
                 }
             }
-        }
 
-        if (expiredList.isEmpty() && expiringSoonList.isEmpty()) return
+        covidCertificateSettings.lastDccStateBackgroundCheck.update { timeStamper.nowUTC }
+    }
 
-        expiredList.forEach {
-            if (!it.hasHadExpiredNotification()) {
-                showNotificationExpired(it)
-                return@forEach
-            }
-        }
-
-        expiringSoonList.forEach {
-            if (!it.hasHadExpiringSoonNotification()) {
-                showNotificationExpiringSoon(it)
-                return@forEach
-            }
+    private suspend fun setStateNotificationShown(certificate: CwaCovidCertificate) {
+        val now = timeStamper.nowUTC
+        val state = certificate.getState()
+        when (certificate) {
+            is RecoveryCertificate -> recoveryRepository.setNotifiedState(certificate.containerId, state, now)
+            is VaccinationCertificate -> vaccinationRepository.setNotifiedState(certificate.containerId, state, now)
+            else -> throw UnsupportedOperationException("Class: ${certificate.javaClass.simpleName}")
         }
     }
 
-    private suspend fun CertificateContainerId.hasHadExpiredNotification(): Boolean {
-        return dccExpirationNotificationRepository.hasExpiredNotificationBeenShownFor(this)
+    private suspend fun getCertificates(): Set<CwaCovidCertificate> {
+        val vacCerts = vaccinationRepository.vaccinationInfos.first().map { it.vaccinationCertificates }.flatten()
+        Timber.tag(TAG).d("Checking %d vaccination certificates", vacCerts.size)
+        val recCerts = recoveryRepository.certificates.first().map { it.recoveryCertificate }
+        Timber.tag(TAG).d("Checking %d recovery certificates", recCerts.size)
+
+        return (vacCerts + recCerts).toSet()
     }
 
-    private suspend fun CertificateContainerId.hasHadExpiringSoonNotification(): Boolean {
-        return dccExpirationNotificationRepository.hasExpiringSoonNotificationBeenShownFor(this)
-    }
-
-    private suspend fun showNotificationExpired(containerId: CertificateContainerId) {
-        dccExpirationNotification.showCheckNotification(containerId)
-        dccExpirationNotificationRepository.reportExpiredNotificationShownFor(containerId)
-    }
-
-    private suspend fun showNotificationExpiringSoon(containerId: CertificateContainerId) {
-        dccExpirationNotification.showCheckNotification(containerId)
-        dccExpirationNotificationRepository.reportExpiringSoonNotificationShownFor(containerId)
+    companion object {
+        private val TAG = DccExpirationCheck::class.java.simpleName
     }
 }
