--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/signature/core/execution/DccStateCheckWorker.kt
@@ -1,33 +1,0 @@
-package de.rki.coronawarnapp.covidcertificate.signature.core.execution
-
-import android.content.Context
-import androidx.work.CoroutineWorker
-import androidx.work.WorkerParameters
-import dagger.assisted.Assisted
-import dagger.assisted.AssistedFactory
-import dagger.assisted.AssistedInject
-import de.rki.coronawarnapp.util.worker.InjectedWorkerFactory
-import timber.log.Timber
-
-class DccStateCheckWorker @AssistedInject constructor(
-    @Assisted val context: Context,
-    @Assisted workerParams: WorkerParameters,
-    private val dccStateCheckWork: DccStateCheckWork,
-) : CoroutineWorker(context, workerParams) {
-
-    override suspend fun doWork(): Result = try {
-        dccStateCheckWork.checkStates()
-
-        Result.success()
-    } catch (e: Exception) {
-        Timber.tag(TAG).e(e, "DCC state check failed.")
-        Result.retry()
-    }
-
-    @AssistedFactory
-    interface Factory : InjectedWorkerFactory<DccStateCheckWorker>
-
-    companion object {
-        private val TAG = DccStateCheckWorker::class.java.simpleName
-    }
-}
